requirements.txt
Django==5.2.7
Pillow==11.3.0
django-unfold==0.43.0
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.2.0